support blogc-make as build tool.
blogc-make is a built tool that replaces make and implements all the
rules needed to build a blogc website. the blogcfile added here covers
all the features previously implemented in the makefile, but the
'reload' task was not implemented in blogc-make yet.

WARNING: blogc-make is still experimental, no stable release includes it
yet! The blogcfile syntax may change!